+++ src/render.cpp
@@ -33,16 +33,16 @@ void render::DrawSprites(const std::vector<entity> &entityList)
 {
 	for(const auto &e : entityList)
 	{
-		const uint16_t type = uint16_t(e.GetEntityType());
+		const uint16_t type = e.GetEntityType();
 		const std::vector<uint16_t> &spriteSheet = spriteSheets[type];
 
 		const EntityState state = e.GetEntityState();
-		const uint8_t &w = spriteRectList[type][uint8_t(state.action)].w;
-		const uint8_t &h = spriteRectList[type][uint8_t(state.action)].h;
+		const uint8_t &w = spriteRectList[type][state.action].w;
+		const uint8_t &h = spriteRectList[type][state.action].h;
 
 		const Position pos = e.GetPosition();
 
-		const std::vector<uint16_t> sprite = SpriteFromSheet(spriteSheet, spriteRectList[type][animationList[type][uint8_t(state.action)][state.animState]], e.GetFlip());
+		const std::vector<uint16_t> sprite = SpriteFromSheet(spriteSheet, spriteRectList[type][animationList[type][state.action][state.animState]], e.GetFlip());
 
 		for(uint16_t y = 0; y < h; ++y)
 		{
